Refactor dashboard and weight analytics components: remove unused goal summary widget, streamline goal display logic, and enhance weight analytics tab with goal tracking features. Update UI elements for improved readability and consistency.
