+++ lib/features/dashboard/home_screen.dart
@@ -1,17 +1,14 @@
 import 'package:caltrack/app/app_snackbar.dart';
 import 'package:caltrack/app/profile_controller.dart';
 import 'package:caltrack/core/food_emoji.dart';
-import 'package:caltrack/core/goal_eta.dart';
 import 'package:caltrack/core/nutrition.dart';
 import 'package:caltrack/core/spacing.dart';
-import 'package:caltrack/core/units.dart';
 import 'package:caltrack/data/caltrack_repository.dart';
 import 'package:caltrack/data/opennutrition_catalog.dart';
 import 'package:caltrack/features/food/food_entry_sheet.dart';
 import 'package:caltrack/features/food/quick_add_sheet.dart';
 import 'package:caltrack/widgets/animated_list_item.dart';
 import 'package:caltrack/widgets/goal_choice_sheet.dart';
-import 'package:caltrack/widgets/goal_editor_sheet.dart';
 import 'package:caltrack/widgets/styled_card.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
@@ -24,7 +21,6 @@ final _fmtDayShort = DateFormat.MMMd();
 final _fmtDayLong = DateFormat.yMMMEd();
 final _fmtTime = DateFormat.jm();
 final _fmtMonthDay = DateFormat.MMMEd();
-final _fmtYmd = DateFormat.yMMMd();
 
 /// Dashboard tab body. Owns the goal-completion celebration sheet and
 /// the scrolling cards. Pure body content — the surrounding shell
@@ -213,7 +209,6 @@ class _DashboardListView extends StatelessWidget {
     final today = calendarDay(DateTime.now());
     final isToday = selectedDay == today;
     const canGoForward = true;
-    final profile = profileCtl.profile!;
 
     return ListView(
       padding: EdgeInsets.fromLTRB(
@@ -283,12 +278,6 @@ class _DashboardListView extends StatelessWidget {
           ),
         ),
         const SizedBox(height: Spacing.md),
-        if (currentGoal != null)
-          AnimatedListItem(
-            index: 3,
-            child: _GoalSummary(goal: currentGoal, profile: profile),
-          ),
-        const SizedBox(height: Spacing.md),
         FutureBuilder<List<WeightEntry>>(
           future: repo.weightEntriesLimit(1),
           builder: (context, wSnap) {
@@ -312,7 +301,7 @@ class _DashboardListView extends StatelessWidget {
                     ),
                     title: const Text('Weigh-in overdue'),
                     subtitle: Text(
-                      'Last entry ${_fmtYmd.format(last.recordedAt)}. '
+                      'Last entry ${_fmtMonthDay.format(last.recordedAt)}. '
                       'Weekly check-ins help adjust your plan.',
                     ),
                     trailing: FilledButton.tonal(
@@ -832,279 +821,6 @@ class _LegendDot extends StatelessWidget {
           label,
           style: theme.textTheme.labelSmall?.copyWith(
             color: theme.colorScheme.onSurfaceVariant,
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _GoalSummary extends StatelessWidget {
-  const _GoalSummary({required this.goal, required this.profile});
-
-  final Goal goal;
-  final Profile profile;
-
-  Future<({double? latestKg, double? trendKgWeek})> _loadDetails(
-    CalTrackRepository repo,
-  ) async {
-    final results = await Future.wait([
-      repo.weightEntriesLimit(1),
-      repo.trendKgPerWeek(windowDays: 14),
-    ]);
-    final entries = results[0] as List<WeightEntry>;
-    final trend = results[1] as double?;
-    return (
-      latestKg: entries.isEmpty ? null : entries.first.weightKg,
-      trendKgWeek: trend,
-    );
-  }
-
-  Future<void> _openEditor(BuildContext context) async {
-    final repo = context.read<CalTrackRepository>();
-    final profile = await repo.requireProfile();
-    if (!context.mounted) return;
-    await showGoalEditorSheet(context, repo: repo, profile: profile);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final scheme = theme.colorScheme;
-    final repo = context.read<CalTrackRepository>();
-    return FutureBuilder<({double? latestKg, double? trendKgWeek})>(
-      future: _loadDetails(repo),
-      builder: (context, snap) {
-        final data = snap.data;
-        if (data == null) return const SizedBox.shrink();
-        final unit = WeightUnit.fromStored(profile.weightUnit);
-        final target = unit == WeightUnit.kg
-            ? '${goal.targetWeightKg.toStringAsFixed(1)} kg'
-            : '${kgToLb(goal.targetWeightKg).toStringAsFixed(1)} lb';
-        final rate = goal.weeklyChangeKgPerWeek;
-        final displayRate = unit == WeightUnit.kg ? rate : kgToLb(rate);
-        String pace;
-        if (goal.status == 'maintain' || rate.abs() < 0.001) {
-          pace = 'Maintenance';
-        } else if (rate < 0) {
-          pace = 'Losing ~${(-displayRate).abs().toStringAsFixed(2)} ${unit.shortLabel}/week';
-        } else {
-          pace = 'Gaining ~${displayRate.toStringAsFixed(2)} ${unit.shortLabel}/week';
-        }
-        final subtitleText =
-            goal.status == 'pending_choice' ? 'Choose next step' : pace;
-
-        return StyledCard(
-          tone: CardTone.low,
-          padding: const EdgeInsets.fromLTRB(
-            Spacing.lg,
-            Spacing.md,
-            Spacing.sm,
-            Spacing.md,
-          ),
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Row(
-                children: [
-                  Container(
-                    width: 36,
-                    height: 36,
-                    decoration: BoxDecoration(
-                      color: scheme.primaryContainer,
-                      borderRadius: Corners.radiusSm,
-                    ),
-                    child: Icon(
-                      Icons.flag_outlined,
-                      size: 20,
-                      color: scheme.onPrimaryContainer,
-                    ),
-                  ),
-                  const SizedBox(width: Spacing.md),
-                  Expanded(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          'Goal: $target',
-                          style: theme.textTheme.titleMedium?.copyWith(
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                        const SizedBox(height: 2),
-                        Text(
-                          subtitleText,
-                          style: theme.textTheme.bodySmall?.copyWith(
-                            color: scheme.onSurfaceVariant,
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-                  IconButton(
-                    tooltip: 'Edit goal',
-                    icon: const Icon(Icons.edit_outlined),
-                    onPressed: () => _openEditor(context),
-                  ),
-                ],
-              ),
-              if (goal.status != 'maintain' && goal.status != 'pending_choice')
-                _GoalEtaSection(
-                  goal: goal,
-                  currentKg: data.latestKg,
-                  trendKgWeek: data.trendKgWeek,
-                  unit: unit,
-                ),
-            ],
-          ),
-        );
-      },
-    );
-  }
-}
-
-/// Static + trend ETA rows under the dashboard goal card. Hidden when
-/// neither prediction is meaningful (no weigh-in, at-goal, etc.).
-class _GoalEtaSection extends StatelessWidget {
-  const _GoalEtaSection({
-    required this.goal,
-    required this.currentKg,
-    required this.trendKgWeek,
-    required this.unit,
-  });
-
-  final Goal goal;
-  final double? currentKg;
-  final double? trendKgWeek;
-  final WeightUnit unit;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final scheme = theme.colorScheme;
-    final current = currentKg;
-    if (current == null) return const SizedBox.shrink();
-
-    final atGoal = atGoalWeight(
-      currentWeightKg: current,
-      targetWeightKg: goal.targetWeightKg,
-    );
-    if (atGoal) {
-      return Padding(
-        padding: const EdgeInsets.only(top: 12, right: 12),
-        child: Text(
-          'You are at your goal weight.',
-          style: theme.textTheme.bodySmall?.copyWith(
-            color: scheme.primary,
-            fontWeight: FontWeight.w600,
-          ),
-        ),
-      );
-    }
-
-    final staticEta = estimateGoalEta(
-      currentWeightKg: current,
-      targetWeightKg: goal.targetWeightKg,
-      weeklyKg: goal.weeklyChangeKgPerWeek,
-    );
-    final trend = trendKgWeek;
-    final trendEta = trend == null
-        ? null
-        : estimateGoalEta(
-            currentWeightKg: current,
-            targetWeightKg: goal.targetWeightKg,
-            weeklyKg: trend,
-          );
-
-    if (staticEta == null && trendEta == null) {
-      return const SizedBox.shrink();
-    }
-
-    return Padding(
-      padding: const EdgeInsets.only(top: 14, right: 12),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          if (staticEta != null)
-            _EtaRow(
-              icon: Icons.timer_outlined,
-              label: 'At your chosen pace',
-              eta: staticEta,
-              color: scheme.primary,
-            ),
-          if (trendEta != null) ...[
-            const SizedBox(height: 6),
-            _EtaRow(
-              icon: Icons.trending_flat,
-              label: 'At your recent trend',
-              eta: trendEta,
-              color: scheme.secondary,
-              trailingNote:
-                  unit == WeightUnit.kg
-                      ? 'based on last ~14 days (${trendEta.weeklyKg.toStringAsFixed(2)} kg/wk)'
-                      : 'based on last ~14 days (${kgToLb(trendEta.weeklyKg).toStringAsFixed(2)} lb/wk)',
-            ),
-          ] else if (staticEta != null)
-            Padding(
-              padding: const EdgeInsets.only(top: 4),
-              child: Text(
-                'Trend ETA appears once you have a few weigh-ins.',
-                style: theme.textTheme.bodySmall?.copyWith(
-                  color: scheme.onSurfaceVariant,
-                ),
-              ),
-            ),
-        ],
-      ),
-    );
-  }
-}
-
-class _EtaRow extends StatelessWidget {
-  const _EtaRow({
-    required this.icon,
-    required this.label,
-    required this.eta,
-    required this.color,
-    this.trailingNote,
-  });
-
-  final IconData icon;
-  final String label;
-  final GoalEta eta;
-  final Color color;
-  final String? trailingNote;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final scheme = theme.colorScheme;
-    final dateLabel = _fmtYmd.format(eta.eta);
-    final weeksLabel =
-        eta.weeks == 1 ? '~1 week' : '~${eta.weeks} weeks';
-    return Row(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Icon(icon, size: 18, color: color),
-        const SizedBox(width: 8),
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                '$label · $weeksLabel',
-                style: theme.textTheme.bodyMedium?.copyWith(
-                  fontWeight: FontWeight.w600,
-                ),
-              ),
-              Text(
-                'around $dateLabel'
-                '${trailingNote == null ? '' : ' · $trailingNote'}',
-                style: theme.textTheme.bodySmall?.copyWith(
-                  color: scheme.onSurfaceVariant,
-                ),
-              ),
-            ],
           ),
         ),
       ],
